apps/worker/pyproject.toml
[project]
name = "arb-worker"
version = "0.1.0"
description = "arb-finder Python worker — scrapers, event matcher, arb recomputation pipeline"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.27",
  "curl-cffi>=0.7",
  "apscheduler>=3.10",
  "tenacity>=9.0",
  "structlog>=24.4",
  "python-dotenv>=1.0",
  "rapidfuzz>=3.9",
  "playwright>=1.48",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "ruff>=0.6"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["arb_worker"]

[project.scripts]
arb-worker = "arb_worker.cli:main"
